+++ TennisClub.Common/Game/GameCreateDTO.cs
@@ -6,7 +6,7 @@ namespace TennisClub.Common.Game
     {
         public string GameNumber { get; set; }
         public int MemberId { get; set; }
-        public byte LeagueId { get; set; }
+        public int LeagueId { get; set; }
         public DateTime Date { get; set; }
     }
 }
