+++ TennisClub.Common/Game/GameReadDTO.cs
@@ -7,7 +7,7 @@ namespace TennisClub.Common.Game
         public int Id { get; set; }
         public string GameNumber { get; set; }
         public int MemberId { get; set; }
-        public byte LeagueId { get; set; }
+        public int LeagueId { get; set; }
         public DateTime Date { get; set; }
     }
 }
