+++ TennisClub.Common/MemberRole/MemberRoleReadDTO.cs
@@ -6,7 +6,7 @@ namespace TennisClub.Common.MemberRole
     {
         public int Id { get; set; }
         public int MemberId { get; set; }
-        public byte RoleId { get; set; }
+        public int RoleId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
     }
